+++ Documentation/com/llevame_app_project/Data/LoginResponseData.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.llevame_app_project.Data.LoginResponseData</li>
+<li>com.llevame_app_project.Data.ResponseData</li>
 </ul>
 </li>
 </ul>
