+++ app/src/main/java/com/llevame_app_project/CardRegistrationActivity.java
@@ -11,8 +11,13 @@ import android.widget.Toast;
 import com.llevame_app_project.Data.PassengerData;
 import com.llevame_app_project.Data.Remote.ApiUtils;
 import com.llevame_app_project.Data.Remote.RegistrationServices;
+import com.llevame_app_project.Data.ResponseData;
 import com.llevame_app_project.Forms.FirstRegistrationForm;
 import com.llevame_app_project.Forms.SecondRegistrationForm;
+
+import java.io.IOException;
+
+import retrofit2.Response;
 
 public class CardRegistrationActivity extends AppCompatActivity {
 
@@ -41,30 +46,42 @@ public class CardRegistrationActivity extends AppCompatActivity {
                 secondForm.lastName = mLastName.getText().toString();
                 secondForm.creditCardNumber = mCreditCardNumber.getText().toString();
 
-                if(firstForm.isDriver) {
+                if (firstForm.isDriver) {
                     intent = new Intent(CardRegistrationActivity.this,
                             CarRegistrationActivity.class);
                     intent.putExtra("firstForm", firstForm);
                     intent.putExtra("secondForm", secondForm);
-                }else {
-                    registerNewPassenger();
-                    intent = new Intent(CardRegistrationActivity.this,
-                            PassengerActivity.class);
-                }
+                    startActivity(intent);
+                } else {
 
-                startActivity(intent);
+                    try {
+                        registerNewPassenger();
+                        intent = new Intent(CardRegistrationActivity.this,
+                                PassengerActivity.class);
+                        startActivity(intent);
+                    } catch (IOException e) {
+                        Toast.makeText(CardRegistrationActivity.this.getBaseContext(),
+                                            "Couldn't reach the server",
+                                            Toast.LENGTH_SHORT).show();
+                    }
+                }
             }
 
-            private void registerNewPassenger(){
+            private void registerNewPassenger() throws IOException {
                 PassengerData passengerData = new PassengerData();
+                Response<ResponseData> response;
+
                 passengerData.setEmail(firstForm.email.toString());
                 passengerData.setCreditCardNumber(secondForm.creditCardNumber.toString());
                 passengerData.setFirstName(secondForm.firstName.toString());
                 passengerData.setLastName(secondForm.lastName.toString());
                 passengerData.setPassword(firstForm.password);
-                ApiUtils.getRegistrationServices().registerUser(passengerData.getEmail(),
-                        passengerData);
+                //TO DO: Check that everything went alright.
+                //Currently server notifies a problem when there isn't
+                response = ApiUtils.getRegistrationServices().registerUser(passengerData.getEmail(),
+                        passengerData).execute();
             }
+
         });
     }
 }
