+++ app/src/main/java/com/llevame_app_project/Data/Remote/LoginServices.java
@@ -1,18 +1,16 @@
 package com.llevame_app_project.Data.Remote;
 
-import com.llevame_app_project.Data.LoginResponseData;
+import com.llevame_app_project.Data.ResponseData;
+
 import retrofit2.Call;
-import retrofit2.http.Body;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
-import retrofit2.http.GET;
-import retrofit2.http.Header;
 import retrofit2.http.PATCH;
 import retrofit2.http.Path;
 
 public interface LoginServices {
     @FormUrlEncoded
     @PATCH("api/v1/account/{userName}")
-    Call<LoginResponseData> loginUser(@Path("userName") String userName,
-                                      @Field("password") String password);
+    Call<ResponseData> loginUser(@Path("userName") String userName,
+                                 @Field("password") String password);
 }
