+++ app/src/main/java/com/llevame_app_project/Data/Remote/PassengerServices.java
@@ -1,7 +1,6 @@
 package com.llevame_app_project.Data.Remote;
 
 import com.llevame_app_project.Data.PassengerData;
-import com.llevame_app_project.Data.UserResponseData;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
