+++ app/src/main/java/com/llevame_app_project/Data/Remote/RegistrationServices.java
@@ -1,16 +1,10 @@
 package com.llevame_app_project.Data.Remote;
 
-import com.llevame_app_project.Data.DriverData;
-import com.llevame_app_project.Data.LoginResponseData;
 import com.llevame_app_project.Data.PassengerData;
-
-import java.sql.Driver;
+import com.llevame_app_project.Data.ResponseData;
 
 import retrofit2.Call;
 import retrofit2.http.Body;
-import retrofit2.http.Field;
-import retrofit2.http.FormUrlEncoded;
-import retrofit2.http.PATCH;
 import retrofit2.http.POST;
 import retrofit2.http.Path;
 
@@ -21,6 +15,6 @@ import retrofit2.http.Path;
 public interface RegistrationServices {
 
     @POST("api/v1/account/{userName}")
-    Call<PassengerData> registerUser(@Path("userName") String userName,
-                                  @Body PassengerData passengerData);
+    Call<ResponseData> registerUser(@Path("userName") String userName,
+                                    @Body PassengerData passengerData);
 }
